alert box functionality

+++ app/assets/javascripts/portal/fulfillment2.js.coffee
@@ -58,14 +58,26 @@ $(document).ready ->
     object_id = $(this).data("#{klass}_id")
     data = {}
     data['study_tracker'] = $('#study_tracker_hidden_field').val() || null
-    $.ajax
-      type: 'DELETE'
-      url:  "/portal/admin/#{klass}s/#{object_id}"
-      data: JSON.stringify(data)
-      dataType: "script"
-      contentType: 'application/json; charset=utf-8'
-      success: ->
-        $().toastmessage('showSuccessToast', "#{klass.humanize()} has been deleted.");
+    confirm_message = "Are you sure that you want to remove this service from all subjects' visit calendars in this arm?"
+    if $(this).data("has_popup") == true
+      if confirm(confirm_message)
+        $.ajax
+          type: 'DELETE'
+          url:  "/portal/admin/#{klass}s/#{object_id}"
+          data: JSON.stringify(data)
+          dataType: "script"
+          contentType: 'application/json; charset=utf-8'
+          success: ->
+            $().toastmessage('showSuccessToast', "#{klass.humanize()} has been deleted.");
+    else
+      $.ajax
+        type: 'DELETE'
+        url:  "/portal/admin/#{klass}s/#{object_id}"
+        data: JSON.stringify(data)
+        dataType: "script"
+        contentType: 'application/json; charset=utf-8'
+        success: ->
+          $().toastmessage('showSuccessToast', "#{klass.humanize()} has been deleted.");
   )
 
   put_attribute = (id, klass, data) ->
@@ -275,14 +287,16 @@ $(document).ready ->
     object_id = $('#delete_ppv_service_id').val()
     data = {}
     data['study_tracker'] = $('#study_tracker_hidden_field').val() || null
-    $.ajax
-      type: 'DELETE'
-      url:  "/portal/admin/line_items/#{object_id}"
-      data: JSON.stringify(data)
-      dataType: "script"
-      contentType: 'application/json; charset=utf-8'
-      success: ->
-        $().toastmessage('showSuccessToast', "#{klass.humanize()} has been deleted.");
+    confirm_message = "Are you sure that you want to remove this service from all subjects' visit calendars?"
+    if confirm(confirm_message)
+      $.ajax
+        type: 'DELETE'
+        url:  "/portal/admin/line_items/#{object_id}"
+        data: JSON.stringify(data)
+        dataType: "script"
+        contentType: 'application/json; charset=utf-8'
+        success: ->
+          $().toastmessage('showSuccessToast', "#{klass.humanize()} has been deleted.");
   )
 
   $(document).on('click', '.expand_li', ->
